cloud_function/requirements.txt
google-cloud-storage==2.10.0
google-cloud-functions==1.12.0
pandas==2.0.3
numpy==1.24.3
flask==2.3.3

